pyproject.toml
[tool.poetry]
name = "on-duty-bot"
version = "0.1.0"
description = ""
authors = ["e.pvlv <egn.pvlv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openpyxl = "^3.1.2"
pytelegrambotapi = "^4.16.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
